skycast_core/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
chainlit
